Move OpType
