+++ automerge-backend/src/helper.rs
@@ -1,7 +1,3 @@
 pub(crate) fn make_true() -> bool {
     true
 }
-
-pub(crate) fn make_false() -> bool {
-    false
-}
