+++ automerge-protocol/src/serde_impls/mod.rs
@@ -5,6 +5,7 @@ mod element_id;
 mod key;
 mod value;
 mod request_key;
+mod op_type;
 
 pub(crate) fn make_false() -> bool {
     return false
